feat(lineMetrics): added source property `lineMetrics`
In order to use line gradients, the option `lineMetrics` is needed. On iOS side, this is achieved by setting MGLShapeSourceOptionLineDistanceMetrics(https://mapbox.github.io/mapbox-gl-native/macos/0.12.0/Other%20Constants.html#/c:@MGLShapeSourceOptionLineDistanceMetrics) and on Android it can be set with `withLineMetrics` on the `GeoJsonOptions`. The option `lineMetrics` was added to the `GeoJSONSource` API object and is used in the `addSource(...)` methods on iOS and Android as described above.