feat(layer type): added property `type` to layer class that returns the GeoJSON layer type
In order to get more information about the type of a layer (e.g. obtained by calling `getLayers`), the GeoJSON layer type was added as property to the layer class. This is for example helpful to find the first symbol layer to add a polyline layer under all map annotations.